refactored store to have seperate files for actions and reducers
