+++ store/actions.js
@@ -1,0 +1,7 @@
+export const actionTypes = {
+    INCREMENT: 'INCREMENT',
+};
+
+export const incrementCount = (isServer) => dispatch => {
+    return dispatch({ type: actionTypes.INCREMENT });
+}
